Interface for RequestStatusEntry
