+++ src/models/configurationSettings.ts
@@ -1,5 +1,6 @@
 import { DocumentWrapper } from "../utils/DocumentWrapper";
 import { MyEventEmitter } from "../utils/myEventEmitter";
+import { RequestStatusEntry } from "../utils/requestStatusEntry";
 import { RequestHeaders } from './base';
 import { FormParamEncodingStrategy } from './formParamEncodingStrategy';
 import { HostCertificate } from './hostCertificate';
@@ -45,4 +46,5 @@ export abstract class RestClientSettings {
     abstract getCurrentHttpFileName(): string | undefined;
     abstract getCurrentDocumentWrapper(): DocumentWrapper | undefined;
     abstract getEmitter<T>(): MyEventEmitter<T>;
+    abstract getRequestStatusEntry(): RequestStatusEntry;
 }
